requirements.txt
pyspark>=3.3.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
boto3>=1.28.0
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0
python-dotenv>=1.0.0